# PythonApi/openapi.json
{"openapi": "3.1.0", "info": {"title": "Product Inventory API", "version": "v1"}, "paths": {"/api/products": {"get": {"summary": "List Products", "operationId": "list_products_api_products_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Product"}, "type": "array", "title": "Response List Products Api Products Get"}}}}}}, "post": {"summary": "Create Product", "operationId": "create_product_api_products_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Response Create Product Api Products Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/products/{id}": {"get": {"summary": "Get Product", "operationId": "get_product_api_products__id__get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Product"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"summary": "Update Product", "operationId": "update_product_api_products__id__put", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Product"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete Product", "operationId": "delete_product_api_products__id__delete", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Product": {"properties": {"name": {"type": "string", "title": "Name"}, "sku": {"type": "string", "title": "Sku"}, "price": {"type": "number", "exclusiveMinimum": 0.0, "title": "Price"}, "stock": {"type": "integer", "minimum": 0.0, "title": "Stock"}, "category": {"type": "string", "title": "Category"}, "id": {"type": "integer", "title": "Id"}}, "type": "object", "required": ["name", "sku", "price", "stock", "category", "id"], "title": "Product"}, "ProductCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "sku": {"type": "string", "title": "Sku"}, "price": {"type": "number", "exclusiveMinimum": 0.0, "title": "Price"}, "stock": {"type": "integer", "minimum": 0.0, "title": "Stock"}, "category": {"type": "string", "title": "Category"}}, "type": "object", "required": ["name", "sku", "price", "stock", "category"], "title": "ProductCreate"}, "ProductUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "sku": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sku"}, "price": {"anyOf": [{"type": "number", "exclusiveMinimum": 0.0}, {"type": "null"}], "title": "Price"}, "stock": {"anyOf": [{"type": "integer", "minimum": 0.0}, {"type": "null"}], "title": "Stock"}, "category": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Category"}}, "type": "object", "title": "ProductUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}